Moved 1 func to `event_filters.py`: filter_relationship_type
