+++ scripts/cat/thoughts.py
@@ -5,9 +5,8 @@ from typing import TYPE_CHECKING
 import i18n
 
 from scripts.cat.enums import CatGroup
-from scripts.events_module.event_filters import event_for_cat
+from scripts.events_module.event_filters import event_for_cat, filter_relationship_type
 from scripts.game_structure.localization import load_lang_resource
-from scripts.utility import filter_relationship_type
 
 if TYPE_CHECKING:
     from scripts.cat.cats import Cat
