+++ tests/test_filter_patrol.py
@@ -9,7 +9,8 @@ from scripts.cat_relations.relationship import Relationship
 from scripts.clan import Clan
 from scripts.events_module.patrol.patrol import PatrolEvent, Patrol
 
-from scripts.utility import filter_relationship_type
+from scripts.events_module.event_filters import filter_relationship_type
+
 
 # TODO: redo them! Filtering is not working like this anymore, but it got removed from .github/workflows/test.yml
 # so they are not failing!
